Update task feature module package.
