+++ common/task/src/main/kotlin/cz/kotox/common/task/data/impl/local/room/DomainTaskDatabase.kt
@@ -1,0 +1,14 @@
+package cz.kotox.common.task.data.impl.local.room
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+
+@Database(
+    entities = [TaskEntity::class, TaskLocalImageEntity::class],
+    version = 1,
+    exportSchema = true,
+)
+//@TypeConverters(Converters::class)
+abstract class DomainTaskDatabase : RoomDatabase() {
+    abstract fun taskDao(): TaskDao
+}
