+++ common/task/src/main/kotlin/cz/kotox/common/task/data/impl/remote/api/TaskApi.kt
@@ -1,0 +1,13 @@
+package cz.kotox.common.task.data.impl.remote.api
+
+import cz.kotox.common.task.data.impl.remote.dto.TaskDTO
+import retrofit2.http.GET
+
+private const val ALL_TASKS_DOCUMENT = "86329e68087a3d2bca54"
+interface TaskApi {
+
+    @GET(ALL_TASKS_DOCUMENT)
+    suspend fun getAllTasks(): List<TaskDTO>
+
+
+}
