--- common/task/src/main/kotlin/cz/kotox/task/data/impl/di/DomainTaskModule.kt
@@ -1,34 +1,0 @@
-package cz.kotox.task.data.impl.di
-
-
-import android.content.Context
-import cz.kotox.core.network.di.CommonRetrofit
-import cz.kotox.task.data.impl.local.room.DomainTaskDatabase
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import retrofit2.Retrofit
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object DomainTaskModule {
-
-
-    val provider = DomainTaskModuleProvider()
-
-    @Provides
-    @Singleton
-    fun provideDomainTaskDatabase(
-        @ApplicationContext context: Context,
-    ): DomainTaskDatabase = provider.provideDomainTaskDatabase(context)
-
-    @Singleton
-    @Provides
-    fun provideTaskDao(db: DomainTaskDatabase) = provider.provideTaskDao(db)
-
-    @Provides
-    fun provideTaskApi(@CommonRetrofit retrofit: Retrofit) = provider.provideTaskApi(retrofit)
-}
