+++ feature/task/src/main/kotlin/cz/kotox/feature/task/detail/ui/TaskDetailActivityIntentUtil.kt
@@ -1,0 +1,18 @@
+package cz.kotox.feature.task.detail.ui
+
+import android.content.Context
+import android.content.Intent
+
+object TaskDetailActivityIntentUtil {
+
+    fun getCameraStartIntent(
+        context: Context,
+        taskId: String
+    ): Intent {
+        val clazz = TaskDetailActivity::class.java
+        return Intent(context.applicationContext, clazz).apply {
+            putExtra(ARG_TASK_ID, taskId)
+        }
+    }
+
+}
