+++ kotox-mobile-task/src/androidTest/kotlin/cz/kotox/task/list/TaskDetailScreenSmokeTest.kt
@@ -4,10 +4,10 @@ import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import cz.kotox.core.ui.theme.KotoxBasicTheme
-import cz.kotox.task.detail.ui.TaskDetailActivity
-import cz.kotox.task.detail.ui.TaskDetailScreen
-import cz.kotox.task.detail.ui.TaskDetailScreenInput
-import cz.kotox.task.detail.ui.UI_TEST_DOWNLOAD_BUTTON_TAG
+import cz.kotox.feature.task.detail.ui.TaskDetailActivity
+import cz.kotox.feature.task.detail.ui.TaskDetailScreen
+import cz.kotox.feature.task.detail.ui.TaskDetailScreenInput
+import cz.kotox.feature.task.detail.ui.UI_TEST_DOWNLOAD_BUTTON_TAG
 import cz.kotox.task.detail.ui.component.TaskSummaryItem
 import cz.kotox.task.domain.api.model.Task
 import dagger.hilt.android.testing.HiltAndroidRule
