+++ task-domain/src/main/kotlin/cz/kotox/task/domain/api/mapper/TaskMapper.kt
@@ -2,9 +2,9 @@ package cz.kotox.task.domain.api.factory
 
 import cz.kotox.core.crypto.CryptoHelper
 import cz.kotox.task.domain.api.model.Task
-import cz.kotox.task.data.impl.local.room.TaskEntity
-import cz.kotox.task.data.impl.remote.dto.TASK_DATE_TIME_PATTERN
-import cz.kotox.task.data.impl.remote.dto.TaskDTO
+import cz.kotox.common.task.data.impl.local.room.TaskEntity
+import cz.kotox.common.task.data.impl.remote.dto.TASK_DATE_TIME_PATTERN
+import cz.kotox.common.task.data.impl.remote.dto.TaskDTO
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
