+++ task-domain/src/main/kotlin/cz/kotox/task/domain/api/usecase/GetOneTaskUseCase.kt
@@ -1,6 +1,6 @@
 package cz.kotox.task.domain.api.usecase
 
-import cz.kotox.task.data.api.respository.TaskRepository
+import cz.kotox.common.task.data.api.respository.TaskRepository
 import cz.kotox.task.domain.api.model.Task
 import cz.kotox.task.domain.api.factory.toTask
 import javax.inject.Inject
