+++ task-domain/src/test/kotlin/cz/kotox/domain/task/impl/remote/api/TaskApiTest.kt
@@ -4,9 +4,9 @@ import cz.kotox.android.core.config.AppProperties
 import cz.kotox.core.network.config.AppNetworkingProperties
 import cz.kotox.core.network.di.NetworkModuleProvider
 import cz.kotox.core.test.MockResponseFileReader
-import cz.kotox.task.data.impl.di.DomainTaskModuleProvider
-import cz.kotox.task.data.impl.remote.dto.TaskDTO
-import cz.kotox.task.data.impl.remote.api.TaskApi
+import cz.kotox.common.task.data.impl.di.DomainTaskModuleProvider
+import cz.kotox.common.task.data.impl.remote.dto.TaskDTO
+import cz.kotox.common.task.data.impl.remote.api.TaskApi
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.test.runTest
 import okhttp3.mockwebserver.MockResponse
